+++ src/sketch.h
@@ -185,7 +185,7 @@ public:
 
     SPolygon                polyLoops;
     SBezierLoopSetSet       bezierLoops;
-    SBezierList             bezierOpens;
+    SBezierLoopSet          bezierOpens;
 
     struct {
         PolyError       how;
@@ -483,6 +483,7 @@ public:
     bool HasEndpoints() const;
     Vector EndpointStart() const;
     Vector EndpointFinish() const;
+    bool IsInPlane(Vector norm, double distance) const;
 
     void RectGetPointsExprs(ExprVector *eap, ExprVector *ebp) const;